pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyQt6 = "6.3.1"
PyQt6-Qt6 = "6.3.1"
PyQt6-sip = "13.4.0"
pyyaml = "6.0"
flask = "2.2.2"
flask-cors = "3.0.10"
requests = "2.28.1"

[tool.poetry.dev-dependencies]
black = "22.8.0"
isort = "5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
